+++ tests/testthat/test-trans-numeric.r
@@ -8,11 +8,27 @@ test_that("Pseudo-log is invertible", {
   )
 })
 
-test_that("Box-Cox is invertible", {
+test_that("Modulus is invertible for negative and positive numbers", {
+  trans <- modulus_trans(p = .1)
+  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+  trans <- modulus_trans(p = -2)
+  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+  trans <- modulus_trans(p = 1)
+  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+})
+
+test_that("Boxcox gives error for negative values", {
   trans <- boxcox_trans(p = .1)
-  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+  expect_error(trans$trans(-10:10))
   trans <- boxcox_trans(p = -2)
-  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+  expect_error(trans$trans(-10:10))
+})
+
+test_that("Boxcox is invertible", {
+  trans <- boxcox_trans(p = .1)
+  expect_equal(trans$inv(trans$trans(0:10)), 0:10)
+  trans <- boxcox_trans(p = -2)
+  expect_equal(trans$inv(trans$trans(0:10)), 0:10)
   trans <- boxcox_trans(p = 1)
-  expect_equal(trans$inv(trans$trans(-10:10)), -10:10)
+  expect_equal(trans$inv(trans$trans(0:10)), 0:10)
 })