+++ NEWS.md
@@ -1,7 +1,9 @@
 # scales 0.5.0.9000
 
-* `boxcox_trans` now implements the modulus transformation, a generalisation
-  appropriate for data with both positive and negative values (@dpseidel, #103).
+* New function `modulus_trans` implements the modulus transformation for positive
+  and negative values (@dpseidel).
+
+* `boxcox_trans` is now invertible for `x >= 0` (@dpseidel, #103).
 
 * New function `time_format()` formats `POSIXt` and `hms` objects (@dpseidel, #88).